@@ -20,21 +20,21 @@
     "@radix-ui/react-slot": "^1.1.0",
     "@radix-ui/react-tabs": "^1.1.0",
     "@radix-ui/react-toggle": "^1.1.0",
-    "@redwoodjs/auth-dbauth-web": "7.7.3",
-    "@redwoodjs/forms": "7.7.3",
-    "@redwoodjs/router": "7.7.3",
-    "@redwoodjs/web": "7.7.3",
-    "@redwoodjs/web-server": "7.7.3",
+    "@redwoodjs/auth-dbauth-web": "8.0.0",
+    "@redwoodjs/forms": "8.0.0",
+    "@redwoodjs/router": "8.0.0",
+    "@redwoodjs/web": "8.0.0",
+    "@redwoodjs/web-server": "8.0.0",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "lexical": "^0.17.0",
     "lucide-react": "^0.427.0",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
     "tailwind-merge": "^2.5.2"
   },
   "devDependencies": {
-    "@redwoodjs/vite": "7.7.3",
+    "@redwoodjs/vite": "8.0.0",
     "@tailwindcss/typography": "^0.5.14",
     "@types/react": "^18.2.55",
     "@types/react-dom": "^18.2.19",
