@@ -1,6 +1,6 @@
 // https://prettier.io/docs/en/options.html
 /** @type {import('prettier').RequiredOptions} */
-module.exports = {
+export default {
   trailingComma: 'es5',
   bracketSpacing: true,
   tabWidth: 2,
@@ -17,5 +17,5 @@ module.exports = {
     },
   ],
   tailwindConfig: './web/config/tailwind.config.js',
-  plugins: [require('prettier-plugin-tailwindcss')],
+  plugins: [await import('prettier-plugin-tailwindcss')],
 }
